Add minutest reporter and remove pry cruft
Pry-rescue was getting a bit annoying and didn’t help much. It’s been
removed but the tests are now run in parallel with a better reporting
format used for output. Pry is still available for debugging through
the Gemfile but you’ll need to require ‘pry’ somewhere to get access to
it. Probably in minitest_helper.rb