Mutators can access the XmlSection instance
This makes it possible for mutators to look at other values already set on the
instance. Instance_mutators that don't need this can use a proc to maintain the
old behaviour of just using the value and any attributes. Because people may
have been using lambdas this is still a breaking change, but this makes the
upgrade less painful than it would be.