Wrap Validators::Base#valid? in a critical section
The code generation in the XML section builder is such that all
instances of a built class share the same underlying validator
objects.

The code in `Validators::Base#valid?` is not thread-safe: there
is a race between the `test` and `message` steps. This can
cause the `value` attribute to have changed in between the call
to `test` and the call to `message`, leading to the wrong message
being output (if the message makes use of the value).

This could be fixed by changing `valid?` to pass the `value`
and `attributes` to the `test` and `message` methods, rather than
storing them as non thread-safe member variables. However, a
quicker fix is simply to wrap `valid?`'s internals in a `Mutex`
to coordinate access to this critical section.