Use org-wide reusable workflow for actionlint
 # What
Previously we added `actionlint` workflows to lots of repos:
- https://github.com/fac/dev-platform/issues/314

They trigger on every push to a PR branch, so most of the time this is unnecessary, as workflows are rarely updated. Really we only need to trigger `actionlint` when workflows are changed.

We're now looking to use org-wide reusable workflows that can be updated in one central place.

 # Note!
The org-wide shared workflow is for private repos only.
We're using this repo as the public mirror for the private workflow since only a few repos are public and require this workflow.
In future we might look to create a `shared-workflows-public` repo if necessary.

Dev-P ticket
 - https://github.com/fac/dev-platform/issues/1002